Online Payment Fraud Detection
The fraudulent transactions can include activities such as phishing scams, account takeover, and money laundering, among others

To predict online payment fraud, we can use machine learning models that are trained on historical data of fraudulent and non-fraudulent transactions. By analyzing various features such as the transaction amount, type of transaction, and the behavior of the user, we can identify patterns and anomalies that are indicative of fraudulent activity.